+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.3.0] - 2024-01-09
 - Add 'is_partial' attribute when rendering partials
 - Add support for strings with ' enclosures
+- Add PartialResponse class to render partials in views
 
 ## [0.2.1] - 2023-12-22
 - Cache 'render_partial_from_template' utility function